--- README.md
@@ -1,7 +1,0 @@
-# Trabalho-Academico
-
-
-Jogo proposto como projeto na materia de Projeto Integrador do curso de Analise e Desenvolvimento de Sistemas na faculdade Serviço Nacional de Aprendizagem comercial.
-Um RPG, por enquanto sem nome.
-
-Para meus colegas de grupo peço apenas que crie sua branch para fazer seu código, faça o commit e faça a pull request após terminar, o restante do grupo irá testar, comentar se aprovaram e iniciarei o merge. Obrigado
